requirements.txt
# Zeabur Worker dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
python-dotenv>=1.0.0

# CrewAI + browser automation (uncomment when ready)
# crewai>=0.28.0
# playwright>=1.40.0
# cloudinary>=1.36.0







